requirements.txt
openai
pandas
pyyaml
